将SofaTraceContext交给spring ioc管理
